front setup for mobile
