+++ src/client/components/Button/index.tsx
@@ -1,15 +1,11 @@
 import Link from 'next/link';
-import styled from 'styled-components';
+import { BtnWrap } from '../../components/Button/styles';
 import PropTypes from 'prop-types';
 
-const Wrap = styled.a`
-  text-decoration: none;
-  letter-spacing: 2px;
-`;
 function Button({ href, name }) {
   return (
     <Link href={href} passHref>
-      <Wrap>{name}</Wrap>
+      <BtnWrap>{name}</BtnWrap>
     </Link>
   );
 }
@@ -18,5 +14,3 @@ Button.propTypes = {
   href: PropTypes.node.isRequired,
   name: PropTypes.node.isRequired,
 };
-
-export default Button;
