--- src/client/components/logo-styled.tsx
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledLogo = styled.div`
-  padding: 0.3rem;
-`;
