--- src/client/components/menu-hamb-styled.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledMenuHamb = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  letter-spacing: 5px;
-`;
