--- src/client/components/menu-hamb.tsx
@@ -1,5 +1,0 @@
-const MenuHamb = () => {
-  return <ul></ul>;
-};
-
-export default MenuHamb;
