--- src/client/components/nav-style.tsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledNav = styled.nav`
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  min-height: 1rem;
-  font-size: 1rem;
-  background: rgb(117, 180, 255);
-  background: linear-gradient(
-    90deg,
-    rgba(117, 180, 255, 1) 0%,
-    rgba(230, 96, 124, 1) 50%,
-    rgba(157, 255, 65, 1) 100%
-  );
-`;
