--- src/client/components/navmenu-styled.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledMenuHamb = styled.section`
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  height: 100%;
-  letter-spacing: 5px;
-`;
